fix:glitches added mvc pattern
