--- nodejs_day6/googleauth/views/auth.ejs
@@ -1,21 +1,0 @@
-<!-- views/auth.ejs -->
-<!doctype html>
-<html>
-<head>
-    <title>Google SignIn</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"> <!-- load bulma css -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css"> <!-- load fontawesome -->
-    <style>
-        body        { padding-top:70px; }
-    </style>
-</head>
-<body>
-<div class="container">
-    <div class="jumbotron text-center text-primary">
-        <h1><span class="fa fa-lock"></span> Social Authentication</h1>
-        <p>Login or Register with:</p>
-        <a href="/auth/google" class="btn btn-danger"><span class="fa fa-google"></span> SignIn with Google</a>
-    </div>
-</div>
-</body>
-</html>
