+++ nodejs_day6/views/index.ejs
@@ -1,23 +1,39 @@
-
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.css"
-    />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Home</title>
-  </head>
-  <body>
-    <h1>Home</h1>
 
-    <% if (typeof message !== 'undefined' ) { %>
-    <p><%= message %></p>
-    <% } %>
+<head>
+  <meta charset="UTF-8" />
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.css" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
+  <!-- load bulma css -->
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
+  <!-- load fontawesome -->
+  <style>
+    body {
+      padding-top: 70px;
+    }
+  </style>
+  <title>Home</title>
+</head>
 
-    <a href="/users/login">Login</a>
-  </body>
+<body>
+  <h1>Home</h1>
+  <div class="container">
+    <div class="jumbotron text-center text-primary">
+      <h1><span class="fa fa-lock"></span> Social Authentication</h1>
+      <p>Login or Register with:</p>
+      <a href="/auth/google" class="btn btn-danger"><span class="fa fa-google"></span> SignIn with Google</a>
+    </div>
+    <% if (typeof message !=='undefined' ) { %>
+      <p>
+        <%= message %>
+      </p>
+      <% } %>
+
+        <a href="/users/login">Login</a>
+</body>
+</div>
+
 </html>
