+++ nodejs_day6/views/login.ejs
@@ -1,61 +1,57 @@
 
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.css"
-    />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Login</title>
-  </head>
-  <body>
-    <h1>Login</h1>
 
-   <ul>
-      <% if (typeof errors != 'undefined') { %> <% errors.forEach(error => { %>
-      <li><%= error.message %></li>
-      <% }) %> <% } %>
-    </ul>
+<head>
+  <meta charset="UTF-8" />
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.css" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+  <title>Login</title>
+</head>
 
-    <ul>
-      <% if(messages.success_msg) { %>
-      <li><%= messages.success_msg %></li>
+<body>
+  <h1>Login</h1>
+
+  <ul>
+    <% if (typeof errors !='undefined' ) { %>
+      <% errors.forEach(error=> { %>
+        <li>
+          <%= error.message %>
+        </li>
+        <% }) %>
+          <% } %>
+  </ul>
+
+  <ul>
+    <% if(messages.success_msg) { %>
+      <li>
+        <%= messages.success_msg %>
+      </li>
       <% } %>
-    </ul>
+  </ul>
 
-    <ul>
-      <% if (messages.error) { %>
-      <li><%= messages.error %></li>
+  <ul>
+    <% if (messages.error) { %>
+      <li>
+        <%= messages.error %>
+      </li>
       <% } %>
-    </ul>
+  </ul>
 
-    <form action="/users/login" method="POST">
-      <div>
-        <input
-          type="email"
-          id="email"
-          name="email"
-          placeholder="Email"
-          required
-        />
-      </div>
-      <div>
-        <input
-          type="password"
-          id="password"
-          name="password"
-          placeholder="Password"
-          required
-        />
-      </div>
-      <div>
-        <input type="submit" value="Login" />
-      </div>
+  <form action="/users/login" method="POST">
+    <div>
+      <input type="email" id="email" name="email" placeholder="Email" required />
+    </div>
+    <div>
+      <input type="password" id="password" name="password" placeholder="Password" required />
+    </div>
+    <div>
+      <input type="submit" value="Login" />
+    </div>
+    <a href="/auth/google">Login with Google</a> <br>
+    <a href="/users/register">Register</a>
+  </form>
+</body>
 
-      <a href="/users/register">Register</a>
-    </form>
-  </body>
 </html>
