+++ nodejs_day6/views/register.ejs
@@ -1,59 +1,48 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.css"
-    />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Register</title>
-  </head>
-  <body>
-    <h1>Register</h1>
 
-    <ul>
-      <% if (typeof errors != 'undefined') { %> <% errors.forEach(error => { %>
-      <li><%= error.message %></li>
-      <% }) %> <% } %>
-    </ul>
-    <form action="/users/register" method="POST">
-      <div>
-        <input type="text" id="username" name="username" placeholder="UserName" required />
-      </div>
-      <div>
-        <input
-          type="email"
-          id="email"
-          name="email"
-          placeholder="Email"
-          required
-        />
-      </div>
-      <div>
-        <input
-          type="password"
-          id="password"
-          name="password"
-          placeholder="Password"
-          required
-        />
-      </div>
-      <div>
-        <input
-          type="password"
-          id="password2"
-          name="password2"
-          placeholder="Confirm Password"
-          required
-        />
-      </div>
-      <div>
-        <input type="submit" value="Register" />
-      </div>
+<head>
+  <meta charset="UTF-8" />
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.css" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+  <title>Register</title>
 
-      <a href="/users/login">Already registered? Login here</a>
-    </form>
-  </body>
+<body>
+  <h1>Register</h1>
+
+  <ul>
+    <% if (typeof errors !='undefined' ) { %>
+      <% errors.forEach(error=> { %>
+        <li>
+          <%= error.message %>
+        </li>
+        <% }) %>
+          <% } %>
+  </ul>
+  <form action="/users/register" method="POST">
+    <div>
+      <input type="text" id="username" name="username" placeholder="UserName" required />
+    </div>
+    <div>
+      <input type="email" id="email" name="email" placeholder="Email" required />
+    </div>
+    <div>
+      <input type="password" id="password" name="password" placeholder="Password" required />
+    </div>
+    <div>
+      <input type="password" id="password2" name="password2" placeholder="Confirm Password" required />
+    </div>
+    <div>
+      <input type="submit" value="Register" />
+    </div>
+
+    <a href="/users/login">Already registered? Login here</a><br>
+    <!-- do you want login with google -->
+    <a href="/auth/google">Login with Google</a>
+
+
+  </form>
+</body>
+
 </html>
